refactor: Move photo interfaces
